+++ FlowChartEditorQt/FlowChart/Creator.cpp
@@ -17,7 +17,7 @@ Creator::Creator() {}
 
 Creator::~Creator() {}
 
-NShape* Creator::Create(Long shapeID, Long x, Long y, Long width, Long height, Long width2, Long height2, char(*contents)) {
+NShape* Creator::Create(Long shapeID, float x, float y, float width, float height, float width2, float height2, char(*contents)) {
 	NShape *shape;
 	if (shapeID == ID_TERMINAL) {
 		shape = new Terminal(x, y, width, height, QColor(255, 153, 153), Qt::SolidLine, QColor(0, 0, 0), contents);
