+++ FlowChartEditorQt/FlowChart/WindowPin.cpp
@@ -1,7 +1,7 @@
 #include "WindowPin.h"
 #include "FlowChartVisitor.h"
 
-WindowPin::WindowPin(Long x, Long y, Long width, Long height, QColor backGroundColor,
+WindowPin::WindowPin(float x, float y, float width, float height, QColor backGroundColor,
 	Qt::PenStyle borderLine, QColor borderColor, String contents)
 	: NShape(x, y, width, height, backGroundColor, borderLine, borderColor, contents) {
 	this->isPinned = true;
